rename ac-imgView into ac-image and improve preview test cases
