+++ ac/ac-image/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ac-imgView",
+  "name": "ac-image",
   "version": "1.0.0",
   "main": "dist/index.js",
   "scripts": {
