+++ ac/ac-image/src/components/ZoomView.jsx
