+++ frog/package.json
@@ -13,7 +13,7 @@
     "ac-form": "1.0.0",
     "ac-iframe": "1.0.0",
     "ac-classifier": "1.0.0",
-    "ac-imgView": "1.0.0",
+    "ac-image": "1.0.0",
     "ac-induction": "1.0.0",
     "ac-prox": "1.0.0",
     "ac-quiz": "1.0.0",
